# specs/api-contract.yaml
openapi: 3.1.0
info:
  title: MCP Hub Lite API
  version: 1.0.0
  description: |
    MCP Hub Lite 轻量级MCP网关系统 REST API 契约。

    ## 概述
    本API契约定义了前端Dashboard与后端网关之间的完整交互规范，包括：
    - 系统健康检查
    - 服务器管理（列表、详情、搜索）
    - CMD标准响应格式
    - 完整的错误处理机制

    ## 协议特性
    - 支持分页查询和过滤
    - 标准CMD响应格式（双轨错误码设计）
    - JSON-RPC 2.0兼容
    - MCP协议特定扩展支持
  contact:
    name: MCP Hub Lite Team
    url: https://github.com/mcp-hub-lite
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: 开发环境
  - url: https://api.mcp-hub-lite.com
    description: 生产环境

# =============================================================================
# 全局标签定义
# =============================================================================
...
parameters:
...
  RequestIdParam:
    in: header
    name: X-Request-ID
    description: 客户端生成的请求ID，用于日志追踪
    required: false
    schema:
      type: string
      format: uuid
    example: "req-550e8400-e29b-41d4-a716-446655440000"

# =============================================================================
# 数据模型定义
# =============================================================================
components:
...
  schemas:
...
    CMDBaseResponse:
      type: object
      description: CMD标准响应格式_基础版
      properties:
        code:
          type: integer
          description: 响应状态码
          example: 200
        message:
          type: string
          description: 响应消息
          example: "success"
        timestamp:
          type: string
          format: date-time
          description: 响应时间戳
          example: "2025-12-16T10:30:00.000Z"
        requestId:
          type: string
          format: uuid
          description: 请求ID，用于追踪
          example: "req-550e8400-e29b-41d4-a716-446655440000"
      required:
        - code
        - message
        - timestamp

    # CMD成功响应
    CMDSuccess:
      type: object
      description: CMD标准响应格式_成功版
      allOf:
        - $ref: '#/components/schemas/CMDBaseResponse'
        - type: object
          properties:
            code:
              type: integer
              enum: [200, 201, 204]
            data:
              type: object
              description: 响应数据载荷
              nullable: true
          required:
            - code
            - data

    # CMD错误响应
    CMDError:
      type: object
      description: CMD标准响应格式_错误版（采用双轨错误码设计）
      allOf:
        - $ref: '#/components/schemas/CMDBaseResponse'
        - type: object
          properties:
            code:
              type: integer
              description: |
                错误码（双轨设计）：
                - 200-299：操作成功
                - 1000-1999：系统级错误（ 인프라故障、运行时错误）
                - 2000-2999：安全错误（认证、授权）
                - 3000-3999：业务错误（业务规则验证）
                - -32000到-32768：MCP标准JSON-RPC错误
                - -32800到-32899：MCP特定错误
                - 6000-6999：MCP Hub Lite网关错误
              example: 3004
            data:
              type: null
              description: 错误响应中data为null
              nullable: false
            error:
              type: object
              description: 详细错误信息
              properties:
                category:
                  type: string
                  enum: [SYSTEM, SECURITY, BUSINESS, API, MCP_PROTOCOL]
                  description: 错误分类
                  example: "BUSINESS"
                severity:
                  type: string
                  enum: [FATAL, ERROR, WARN, INFO]
                  description: 错误严重程度
                  example: "ERROR"
                stack:
                  type: string
                  description: 错误堆栈（开发环境提供）
                  example: "Error: 验证失败\n    at..."
                  nullable: true
                context:
                  type: object
                  description: 错误上下文信息
                  additionalProperties: true
                  example:
                    resourceId: "server-001"
                    operation: "fetch"
                  nullable: true
                httpStatus:
                  type: integer
                  description: HTTP状态码（兼容性）
                  example: 404
          required:
            - category
            - severity

    # ------------------
    # 服务器相关模型
    # ------------------

    # 服务器状态枚举
    ServerStatus:
      type: string
      enum: [online, offline, error, starting, stopping]
      description: MCP服务器状态
      example: "online"

    # 服务器统计信息
    ServerStatistics:
      type: object
      description: MCP服务器性能统计
      properties:
        latencyMs:
          type: number
          format: float
          minimum: 0
          description: 平均延迟（毫秒）
          example: 45.6
        toolCount:
          type: integer
          minimum: 0
          description: 可用工具数量
          example: 15
        uptimeSeconds:
          type: integer
          minimum: 0
          description: 运行时间（秒）
          example: 86400
        lastCallTimestamp:
          type: string
          format: date-time
          description: 最后一次调用时间
          example: "2025-12-16T11:29:59.000Z"
      required:
        - latencyMs
        - toolCount
        - uptimeSeconds

    # MCP服务器模型
    Server:
      type: object
      description: MCP服务器信息模型
      properties:
        id:
          type: string
          format: uuid
          description: 服务器唯一标识符
          example: "srv-550e8400-e29b-41d4-a716-446655440000"
        name:
          type: string
          description: 服务器显示名称
          minLength: 1
          maxLength: 100
          example: "MySQL数据库工具"
        description:
          type: string
          description: 服务器详细描述
          maxLength: 500
          nullable: true
          example: "提供MySQL数据库查询和管理功能"
        status:
          $ref: '#/components/schemas/ServerStatus'
        tags:
          type: object
          description: 服务器标签（键值对形式）
          additionalProperties:
            type: string
          example:
            category: "database"
            environment: "production"
            version: "8.0"
        lastSeen:
          type: string
          format: date-time
          description: 最后活跃时间
          example: "2025-12-16T10:29:59.000Z"
        statistics:
          $ref: '#/components/schemas/ServerStatistics'
          nullable: true
        endpoint:
          type: string
          format: uri
          description: MCP服务器连接端点
          example: "stdio://localhost:3001"
          nullable: true
        version:
          type: string
          description: MCP服务器版本
          maxLength: 50
          example: "1.2.3"
          nullable: true
        mcpVersion:
          type: string
          description: 支持的MCP协议版本
          maxLength: 20
          example: "2025-11-25"
          nullable: true
        createdAt:
          type: string
          format: date-time
          description: 创建时间
          example: "2025-12-15T08:00:00.000Z"
        updatedAt:
          type: string
          format: date-time
          description: 更新时间
          example: "2025-12-16T10:30:00.000Z"
      required:
        - id
        - name
        - status
        - lastSeen
        - createdAt
        - updatedAt

    # 服务器列表响应
...
    ServerDetailResponse:
      type: object
      description: 获取服务器详情响应
      allOf:
        - $ref: '#/components/schemas/CMDSuccess'
        - type: object
          properties:
            data:
              allOf:
                - $ref: '#/components/schemas/Server'
                - type: object
                  properties:
                    # 详情中可能包含的额外信息
                    detailedMetrics:
                      type: object
                      description: 详细指标信息
                      additionalProperties: true
                      nullable: true
                    recentCalls:
                      type: array
                      items:
                        type: object
                        properties:
                          timestamp:
                            type: string
                            format: date-time
                          method:
                            type: string
                          durationMs:
                            type: number
                      description: 最近调用记录
                      nullable: true

    # ------------------
    # 搜索相关模型
    # ------------------

    # 搜索查询参数
...
    ErrorResponse:
      type: object
      description: 标准错误响应格式示例
      allOf:
        - $ref: '#/components/schemas/CMDError'
      example:
        code: 3004
        message: "服务器不存在"
        data: null
        timestamp: "2025-12-16T10:30:00.000Z"
        requestId: "req-550e8400-e29b-41d4-a716-446655440000"
        error:
          category: "BUSINESS"
          severity: "ERROR"
          context:
            serverId: "srv-550e8400-e29b-41d4-a716-446655440000"

# =============================================================================
# API端点定义
# =============================================================================
paths:
...
  /api/servers/{serverId}:
    get:
      tags:
        - Servers
      summary: 获取服务器详情
      description: |
        根据服务器ID获取详细的MCP服务器信息。

        返回的详情信息包括：
        - 服务器基本信息和配置
        - 实时性能统计指标
        - 最近调用记录（如果有）
        - 详细的技术栈信息

        **注意**：此接口返回比列表接口更详细的数据。
      operationId: getServerById
      parameters:
        - name: serverId
          in: path
          description: 服务器唯一标识符
          required: true
          schema:
            type: string
            format: uuid
          example: "srv-550e8400-e29b-41d4-a716-446655440000"
        - $ref: '#/parameters/RequestIdParam'
      responses:
        '200':
          description: 成功获取服务器详情
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServerDetailResponse'
              example:
                code: 200
                message: "获取服务器详情成功"
                data:
                  id: "srv-550e8400-e29b-41d4-a716-446655440000"
                  name: "MySQL数据库工具"
                  description: "提供MySQL数据库查询和管理功能，包括CRUD操作、数据备份和性能监控"
                  status: "online"
                  tags:
                    category: "database"
                    environment: "production"
                    version: "8.0"
                  lastSeen: "2025-12-16T11:29:59.000Z"
                  statistics:
                    latencyMs: 45.6
                    toolCount: 15
                    uptimeSeconds: 86400
                    lastCallTimestamp: "2025-12-16T11:29:59.000Z"
                  endpoint: "stdio://localhost:3001"
                  version: "1.2.3"
                  mcpVersion: "2025-11-25"
                  createdAt: "2025-12-15T08:00:00.000Z"
                  updatedAt: "2025-12-16T10:30:00.000Z"
                  detailedMetrics:
                    cpuUsage: 15.3
                    memoryUsage: 128.5
                    connectionCount: 5
                  recentCalls:
                    - timestamp: "2025-12-16T11:15:30.000Z"
                      method: "tools/call"
                      durationMs: 45
                    - timestamp: "2025-12-16T11:10:15.000Z"
                      method: "tools/list"
                      durationMs: 12
                timestamp: "2025-12-16T10:30:00.000Z"
                requestId: "req-550e8400-e29b-41d4-a716-446655440000"
        '404':
          description: 服务器不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 3004
                message: "服务器不存在"
                data: null
                timestamp: "2025-12-16T10:30:00.000Z"
                requestId: "req-550e8400-e29b-41d4-a716-446655440000"
                error:
                  category: "BUSINESS"
                  severity: "ERROR"
                  context:
                    serverId: "srv-550e8400-e29b-41d4-a716-446655440000"
                    availableIds: ["srv-001", "srv-002"]
                  httpStatus: 404

  # ----------------------------------------------------------------------
  # 搜索端点
  # ----------------------------------------------------------------------
...
  /api/servers:
    post:
      tags:
        - Servers
      summary: 添加服务器（已预留，当前版本暂未实现）
      description: |
        添加新的MCP服务器配置。

        **注意**：此接口已在契约中预留，但当前MVP版本暂不实现。
        计划在后续版本中提供以下功能：
        - 动态添加MCP服务器配置
        - 服务器配置验证
        - 批量导入支持
        - 配置模板功能

        当前版本如需添加服务器，请参考系统配置文件：
        `config/servers.json`
      operationId: addServer
      deprecated: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: 服务器配置（预留）
              example:
                name: "新增数据库工具"
                description: "预留接口，暂未实现"
                endpoint: "stdio://localhost:3002"
                tags:
                  category: "database"
      responses:
        '501':
          description: 功能尚未实现
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                code: 5012
                message: "添加服务器功能尚未在当前版本中实现"
                data: null
                timestamp: "2025-12-16T10:30:00.000Z"
                requestId: "req-550e8400-e29b-41d4-a716-446655440000"
                error:
                  category: "BUSINESS"
                  severity: "INFO"
                  context:
                    hint: "请参考 config/servers.json 手动配置"
                  httpStatus: 501

# =============================================================================
# 扩展配置
# =============================================================================

# =============================================================================
# 安全配置
# =============================================================================

# 当前版本暂不需要认证，如果需要可以开启以下配置：
# security:
#   - BearerAuth: []

# components:
#   securitySchemes:
#     BearerAuth:
#       type: http
#       scheme: bearer
#       bearerFormat: JWT
#       description: JWT认证令牌（预留，当前版本未启用）

# =============================================================================
# 响应头定义
# =============================================================================

# components:
#   headers:
#     X-Total-Count:
#       description: 总记录数
#       schema:
#         type: integer
#         example: 100
#     X-Page:
#       description: 当前页码
#       schema:
#         type: integer
#         example: 1
#     X-Limit:
#       description: 每页数量
#       schema:
#         type: integer
#         example: 20

# =============================================================================
# 回调定义（预留）
# # =============================================================================

# 当前版本不涉及 webhook/回调功能，如需可在此处定义

# =============================================================================
# 外部文档链接
# =============================================================================

# externalDocs:
#   description: MCP Hub Lite API 完整文档
#   url: https://docs.mcp-hub-lite.com/